# Parallelize provider installation