--- a/internal/command/init.go
+++ b/internal/command/init.go
@@ -960,6 +960,8 @@ func (c *InitCommand) getProviders(ctx context.Context, config *configs.Config, 
 			}
 		},
 	}
+	// Ensure that events emitted on multiple routines do not trigger race conditions
+	evts = evts.Serialized()
 	ctx = evts.OnContext(ctx)
 
 	mode := providercache.InstallNewProvidersOnly