Now that https://github.com/opentofu/opentofu/pull/2726 is merged, we can safely parallelize the provider installation process.

This PR makes two direct changes:
* Fetch each provider version metadata in it's  own go routine
* Fetch each provider version binary in it's own go routine

And one secondary change: Modify the memoize source to be more granular.  This was necessary to allow the metadata fetching to happen in parallel.  Discovering this was helped with @Yantrio's recent OTel tracing PRs :+1: 

Considerations:
* Go channels paired with result structs were used in this PR.  It could easily be adapted to sync.WaitGroup and locking instead.
* This will now send of lots of requests to the provider source (meta and download), which we may want to consider a rate limit for.  I haven't hit this (yet) in testing, but it is worth considering.
* Each commit message is well formatted so this PR should be squashed

Anecdata:
* Cold cache `tofu init` with 5 large providers: 13s -> 7s (depending on network throughput)
* Warm cache `tofu init` with 5 large providers: 2.2s -> 1.3s

## Target Release

1.10.0

## Checklist

<!-- Please check of ALL items in this list for all PRs: -->

- [x] I have read the [contribution guide](https://github.com/opentofu/opentofu/blob/main/CONTRIBUTING.md).
- [x] I have not used an AI coding assistant to create this PR.
- [x] I have written all code in this PR myself OR I have marked all code I have not written myself (including modified code, e.g. copied from other places and then modified) with a comment indicating where it came from.
- [x] I (and other contributors to this PR) have not looked at the Terraform source code while implementing this PR.

### Go checklist

<!-- If your PR contains Go code, please make sure you check off all items on this list: --> 

- [x] I have run golangci-lint on my change and receive no errors relevant to my code.
- [x] I have run existing tests to ensure my code doesn't break anything.
- [x] I have added tests for all relevant use cases of my code, and those tests are passing.
- [x] I have only exported functions, variables and structs that should be used from other packages.
- [x] I have added meaningful comments to all exported functions, variables, and structs.

### Website/documentation checklist

<!-- If you have changed the website, please follow this checklist: -->

- [ ] I have locally started the website as [described here](https://github.com/opentofu/opentofu/blob/main/website/README.md) and checked my changes.