@apparentlymart 2025-05-13
In case you didn't notice, the race detector seems to think this introduces a new data race. :grimacing: